Added possibility of overriding PGF inclusion command
* ox-latex.el (org-latex--inline-image): The user can now customize the
  way PGF images are included by changing the (new) variable
  `org-latex-inline-pgf-command'.

If the PGF file is stored in a subdirectory and references an external
image file in the same directory, LaTeX will complain about not finding
it.  A workaround could be to define a function like this in the
preamble:

\newcommand\inputpgf[2]{{
    \let\pgfimageWithoutPath\pgfimage
    \renewcommand{\pgfimage}[2][]{\pgfimageWithoutPath[##1]{#1/##2}}
    \input{#1/#2}
  }}

and customizing `org-latex-inline-pgf-command' to hold this function:

(lambda (path)
  (let ((dir-name (substring (file-name-directory path) 0 -1))
        (file-name (file-name-nondirectory path)))
    (format "\\inputpgf{%s}{%s}" dir-name file-name)))

This way, LaTeX will be able to find the external images.  The idea came
from here: http://tex.stackexchange.com/a/282110/9742

TINYCHANGE